refactor the BaseOptimWrapper
